back-end/requirements.txt
timedelta==2020.12.3
asgiref==3.7.2
Django==4.2.10
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-filter==0.2.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.10.0
asgiref==3.7.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
timedelta==2020.12.3
